// frontend/src/api/models/TicketCredential.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proof Pass API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface TicketCredential
 */
export interface TicketCredential {
    /**
     *
     * @type {string}
     * @memberof TicketCredential
     */
    id?: string;
    /**
     *
     * @type {string}
     * @memberof TicketCredential
     */
    email?: string;
    /**
     *
     * @type {string}
     * @memberof TicketCredential
     */
    eventId?: string;
    /**
     *
     * @type {string}
     * @memberof TicketCredential
     */
    data?: string;
    /**
     *
     * @type {Date}
     * @memberof TicketCredential
     */
    issuedAt?: Date;
    /**
     *
     * @type {Date}
     * @memberof TicketCredential
     */
    expireAt?: Date;
}

/**
 * Check if a given object implements the TicketCredential interface.
 */
export function instanceOfTicketCredential(
    value: object,
): value is TicketCredential {
    return true;
}

export function TicketCredentialFromJSON(json: any): TicketCredential {
    return TicketCredentialFromJSONTyped(json, false);
}

export function TicketCredentialFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): TicketCredential {
    if (json == null) {
        return json;
    }
    return {
        id: json['id'] == null ? undefined : json['id'],
        email: json['email'] == null ? undefined : json['email'],
        eventId: json['event_id'] == null ? undefined : json['event_id'],
        data: json['data'] == null ? undefined : json['data'],
        issuedAt:
            json['issued_at'] == null ? undefined : new Date(json['issued_at']),
        expireAt:
            json['expire_at'] == null ? undefined : new Date(json['expire_at']),
    };
}

export function TicketCredentialToJSON(value?: TicketCredential | null): any {
    if (value == null) {
        return value;
    }
    return {
        id: value['id'],
        email: value['email'],
        event_id: value['eventId'],
        data: value['data'],
        issued_at:
            value['issuedAt'] == null
                ? undefined
                : value['issuedAt'].toISOString(),
        expire_at:
            value['expireAt'] == null
                ? undefined
                : value['expireAt'].toISOString(),
    };
}